+++ Orbitals/HolyLight_Proj.cs
@@ -3,6 +3,7 @@ using Microsoft.Xna.Framework;
 using Terraria;
 using Terraria.ID;
 using Terraria.ModLoader;
+using static Virtuous.Tools;
 
 namespace Virtuous.Orbitals
 {
@@ -17,7 +18,7 @@ namespace Virtuous.Orbitals
         private const float DistanceAvg = 70f; //Average distance to the player
         private const float OscSpeedMax = 0.2f; //How fast it can oscillate in and out, which is to say, how far it can go before changing direction
         private const float OscAcc = OscSpeedMax / 60; //How quickly it changes oscillation speed, which is to say, how quickly it reaches the point of direction change
-        private const float OrbSpeed = 1/30f * Tools.RevolutionPerSecond; //Orbiting speed
+        private const float OrbSpeed = 1/30f * RevolutionPerSecond; //Orbiting speed
         private const float RotSpeed = -OrbSpeed; //How fast it rotates the sprite
         private bool  outwards { get { return projectile.ai[1] == 0; } set { projectile.ai[1] = value ? 0 : 1; } } //Direction it's currently oscillating in. Stores into the projectile's built-in ai[1], which is 0 by default (true in this case)
         private float distance = DistanceAvg; //Distance away from the player
@@ -49,10 +50,13 @@ namespace Virtuous.Orbitals
 
         public override void AI()
         {
+            //if (projectile.owner != Main.myPlayer) return; //Only runs AI for the client
+			projectile.netUpdate = true;
+			
             Player player = Main.player[projectile.owner];
             OrbitalPlayer orbitalPlayer = player.GetModPlayer<OrbitalPlayer>();
 
-            if (!orbitalPlayer.active[OrbitalID.HolyLight]) //Keep it alive only while the summon is active
+            if (!orbitalPlayer.active[OrbitalID.HolyLight] && projectile.owner == Main.myPlayer) //Keep it alive only while the summon is active
             {
                 projectile.Kill();
             }
@@ -63,7 +67,7 @@ namespace Virtuous.Orbitals
                     firstTick = false;
                     //The projectile's desired rotation and position was passed as its velocity, so we utilize it then set it to 0 so it doesn't move
                     projectile.rotation = projectile.velocity.ToRotation();
-                    relativePos = projectile.velocity.SafeNormalize(Vector2.UnitY) * distance;
+                    relativePos = projectile.velocity.OfLength(distance);
                     projectile.velocity = Vector2.Zero;
 
                     projectile.Center = player.MountedCenter + relativePos;
@@ -83,7 +87,7 @@ namespace Virtuous.Orbitals
                 distance += oscillationSpeed;
 
                 //Makes it orbit the player
-                relativePos = relativePos.RotatedBy(OrbSpeed).SafeNormalize(Vector2.UnitX) * distance; //Rotates the projectile around the player and resets its distance
+                relativePos = relativePos.RotatedBy(OrbSpeed).OfLength(distance); //Rotates the projectile around the player and resets its distance
                 projectile.rotation += RotSpeed; //Rotates the projectile itself
                 projectile.Center = player.MountedCenter + relativePos; //Moves the projectile to the new position around the player
