+++ Orbitals/Shuriken_Proj.cs
@@ -3,6 +3,7 @@ using Microsoft.Xna.Framework;
 using Terraria;
 using Terraria.ID;
 using Terraria.ModLoader;
+using static Virtuous.Tools;
 
 namespace Virtuous.Orbitals
 {
@@ -18,12 +19,12 @@ namespace Virtuous.Orbitals
         private const float DistanceAvg = 120f; //Average distance to the player when passive
         private const float OscSpeedMax = 6.0f; //How fast it can oscillate in and out, which is to say, how far it can go before changing direction
         private const float OscAcc = OscSpeedMax / 20; //How quickly it changes oscillation speed, which is to say, how quickly it reaches the point of direction change
-        private const float OrbSpeed = 1 * Tools.RevolutionPerSecond; //Orbiting speed
+        private const float OrbSpeed = 1 * RevolutionPerSecond; //Orbiting speed
         private bool outwards { get { return projectile.ai[1] == 0; } set { projectile.ai[1] = value ? 0 : 1; } } //Direction it's currently moving in (away or into the player). Stores into the projectile's built-in ai[1], which is 0 by default (true in this case)
 
         //Final spin
         public  const int DyingTime = 30; //Time it spends doing the final spin
-        private const float DyingOrbSpeed = 4 * Tools.RevolutionPerSecond; //Final spin speed
+        private const float DyingOrbSpeed = 4 * RevolutionPerSecond; //Final spin speed
 
         //Main
         private bool firstTick { get { return projectile.ai[0] == 0; } set { projectile.ai[0] = value ? 0 : 1; } } //Stores into the projectile's built-in ai[0], which is 0 by default (true in this case)
@@ -48,10 +49,13 @@ namespace Virtuous.Orbitals
 
         public override void AI()
         {
+            //if (projectile.owner != Main.myPlayer) return; //Only runs AI for the client
+			projectile.netUpdate = true;
+
             Player player = Main.player[projectile.owner];
             OrbitalPlayer orbitalPlayer = player.GetModPlayer<OrbitalPlayer>();
 
-            if (!orbitalPlayer.active[OrbitalID.Shuriken]) //Keep it alive only while the summon is active
+            if (!orbitalPlayer.active[OrbitalID.Shuriken] && projectile.owner == Main.myPlayer) //Keep it alive only while the summon is active
             {
                 projectile.Kill();
             }
@@ -62,7 +66,7 @@ namespace Virtuous.Orbitals
                     firstTick = false;
                     //The projectile's desired rotation was passed as its velocity, so we utilize it then set it to 0 so it doesn't move
                     projectile.rotation = projectile.velocity.ToRotation();
-                    relativePos = projectile.velocity.SafeNormalize(Vector2.UnitY) * distance;
+                    relativePos = projectile.velocity.OfLength(distance);
                     projectile.velocity = Vector2.Zero;
 
                     //projectile.Center = player.MountedCenter + relativePos;
@@ -76,7 +80,7 @@ namespace Virtuous.Orbitals
                 }
 
                 //Before it spins and dies
-                if (orbitalPlayer.time > DyingTime)
+                if (orbitalPlayer.time > DyingTime || projectile.owner != Main.myPlayer)
                 {
                     projectile.timeLeft = DyingTime; //Keep it from dying naturally
 
@@ -108,7 +112,7 @@ namespace Virtuous.Orbitals
                 float newOrbitingSpeed = OrbSpeed; //Actual orbiting speed used
                 if (orbitalPlayer.time <= DyingTime) newOrbitingSpeed = DyingOrbSpeed; //Changes it if it's the final spin
                 //Makes them orbit the player
-                relativePos = relativePos.RotatedBy(newOrbitingSpeed).SafeNormalize(Vector2.UnitX) * distance; //Rotates the shuriken around the player and resets its distance
+                relativePos = relativePos.RotatedBy(newOrbitingSpeed).OfLength(distance); //Rotates the shuriken around the player and resets its distance
                 projectile.rotation += -newOrbitingSpeed * 2.0f; //Rotates the sprite as well
                 projectile.Center = player.MountedCenter + relativePos; //Moves the shuriken to the defined position around the player
