+++ Orbitals/SpiralSword_Proj.cs
@@ -2,6 +2,7 @@
 using Microsoft.Xna.Framework;
 using Terraria;
 using Terraria.ModLoader;
+using static Virtuous.Tools;
 
 namespace Virtuous.Orbitals
 {
@@ -29,7 +30,7 @@ namespace Virtuous.Orbitals
         private float shootSpeed = 15; //Speed when it's flying outward
 
         //Main
-        private const float AngularSpeed = 1/2f * Tools.RevolutionPerSecond; //How fast it orbits around the player
+        private const float AngularSpeed = 1/2f * RevolutionPerSecond; //How fast it orbits around the player
         private bool firstTick { get { return projectile.ai[0] == 0; } set { projectile.ai[0] = value ? 0 : 1; } } //Stores into the projectile's built-in ai[0], which is 0 by default (true in this case)
         private bool outwards { get { return projectile.ai[1] == 0; } set { projectile.ai[1] = value ? 0 : 1; } } //Direction it's currently moving in (away or into the player). Stores into the projectile's built-in ai[1], which is 0 by default (true in this case)
         private Vector2 relativePos; //Relative position to the player
@@ -40,8 +41,8 @@ namespace Virtuous.Orbitals
 
         public override void SetDefaults()
         {
-            projectile.width = 70;
-            projectile.height = 70;
+            projectile.width = 76;
+            projectile.height = 76;
             projectile.alpha = 50;
             projectile.timeLeft = DyingTime;
             projectile.penetrate = -1;
@@ -54,10 +55,13 @@ namespace Virtuous.Orbitals
 
         public override void AI()
         {
+            //if (projectile.owner != Main.myPlayer) return; //Only runs AI for the client
+			projectile.netUpdate = true;
+
             Player player = Main.player[projectile.owner];
             OrbitalPlayer orbitalPlayer = player.GetModPlayer<OrbitalPlayer>();
 
-            if (!orbitalPlayer.active[OrbitalID.SpiralSword]) //Keep it alive only while the summon is active
+            if (!orbitalPlayer.active[OrbitalID.SpiralSword] && projectile.owner == Main.myPlayer) //Keep it alive only while the summon is active
             {
                 projectile.Kill();
             }
@@ -67,14 +71,14 @@ namespace Virtuous.Orbitals
                 {
                     //The projectile's desired rotation was passed as its velocity, so we utilize it then set it to 0 so it doesn't move
                     projectile.rotation = projectile.velocity.ToRotation() + 45.ToRadians(); //45 degrees because of the sprite
-                    relativePos = projectile.velocity.SafeNormalize(Vector2.UnitY) * distance;
+                    relativePos = projectile.velocity.OfLength(distance);
                     projectile.velocity = Vector2.Zero;
                     originalDamage = projectile.damage;
                     firstTick = false;
                 }
 
                 //Before it shoots out and dies
-                if (orbitalPlayer.time > DyingTime)
+                if (orbitalPlayer.time > DyingTime || projectile.owner != Main.myPlayer)
                 {
                     projectile.timeLeft = DyingTime; //Keep it from dying naturally
 
@@ -116,7 +120,7 @@ namespace Virtuous.Orbitals
                     //Makes them orbit the player
                     float newAngSpeed = AngularSpeed; //Actual angular speed used
                     if (rightClickTimer != 0 && rightClickTimer <= 30) newAngSpeed *= 2; //Double speed the first 30 ticks of shooting out
-                    relativePos = relativePos.RotatedBy(newAngSpeed).SafeNormalize(Vector2.UnitX) * distance; //Rotates the sword around the player and resets its distance
+                    relativePos = relativePos.RotatedBy(newAngSpeed).OfLength(distance); //Rotates the sword around the player and resets its distance
                     projectile.rotation += newAngSpeed; //Rotates the sprite accordingly
                     projectile.Center = player.MountedCenter + relativePos; //Moves the sword to the defined position around the player
 
@@ -124,7 +128,7 @@ namespace Virtuous.Orbitals
                 else //timeLeft actually starts going down from dyingTime while it flies away
                 {
                     projectile.damage = originalDamage * 5; //Deals more damage when shooting out
-                    projectile.Center += relativePos.SafeNormalize(Vector2.UnitY) * shootSpeed; //Projectile moves forward in the direction of its rotation
+                    projectile.Center += relativePos.OfLength(shootSpeed); //Projectile moves forward in the direction of its rotation
                     shootSpeed += ShootAcc; //Accelerate
                 }
             }
@@ -137,7 +141,7 @@ namespace Virtuous.Orbitals
 
         public override Color? GetAlpha(Color newColor) //Fullbright
         {
-            return new Color(28, 255, 255, 150) * projectile.Opacity;
+            return new Color(150, 255, 230, 150) * projectile.Opacity;
         }
     }
 }
